+++ resources/views/dashboard/Role/index.blade.php
@@ -81,6 +81,7 @@
                                                     title="" data-original-title="Settings">
                                                     <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-eye text-primary"><path d="M1 12s4-8 11-8 11 8 11 8-4 8-11 8-11-8-11-8z"></path><circle cx="12" cy="12" r="3"></circle></svg>
                                                 </a> --}}
+                                                @if (Auth::user()->hasPermission('app.edit-role'))
                                                 <a href="{{ route('roles.edit', $role) }}" data-toggle="tooltip" data-placement="top"
                                                     title="" data-original-title="Edit"><svg
                                                         xmlns="http://www.w3.org/2000/svg" width="24" height="24"
@@ -89,8 +90,10 @@
                                                         class="feather feather-edit-2 text-success">
                                                         <path d="M17 3a2.828 2.828 0 1 1 4 4L7.5 20.5 2 22l1.5-5.5L17 3z">
                                                         </path>
-                                                    </svg></a>
-                                                @if ($role->deleteable == true)
+                                                    </svg>
+                                                </a>
+                                                @endif
+                                                @if ($role->deleteable == true && Auth::user()->hasPermission('app.delete-role'))
                                                     <button type="button" onclick="deleteData({{ $role->id }})">
                                                         <svg
                                                             xmlns="http://www.w3.org/2000/svg" width="24" height="24"
